Extract internal/connectcore engine from desktop/vpnservice (ADR-001 A1)
Behavior-preserving refactor: the UI-agnostic client connection engine —
connect state machine, candidate ladder + ranking, WSS front fallback,
punch attempt, health monitor/failover, directory cache, and probes —
moves from desktop/vpnservice into the root package internal/connectcore.
desktop/vpnservice stays behind as a thin Wails adapter: event emission,
the log ring and its 5 Hz coalescing, and desktop/persist +
desktop/proxymode + desktop/proxyconfig wiring behind the engine's new
platform interfaces (EventSink, Persistence, OSProxy, Elevation, and a
ResolveProxyPort seam).

Along with the engine, desktop/discovery moves to internal/discovery, the
engine tuning constants move from desktop/config into connectcore (the
now-empty package is deleted; its SourceURL was unused — the frontend has
its own copy), and the generic proxy-endpoint primitives (loopback host,
OPENRUNG_PROXY_PORT, bind check, per-process pinning) move from
desktop/proxyconfig into connectcore, with proxyconfig re-exporting the
constants. persist.PrependRecent (recents policy) moves into the engine.

Moved tests move with the code; connectcore must not import anything
under desktop/ (enforced by the module graph: desktop is a separate
module that depends on the root).
