+++ .github/workflows/go-checks.yml
@@ -115,9 +115,9 @@ jobs:
       - name: desktop module (vet + test + build)
         working-directory: desktop
         run: |
-          go vet ./vpnservice/... ./config/... ./discovery/... ./persist/... ./proxyconfig/... ./proxymode/...
-          go test -race ./vpnservice/... ./config/... ./discovery/... ./persist/... ./proxyconfig/... ./proxymode/...
-          go build ./vpnservice/... ./config/... ./discovery/... ./persist/... ./proxyconfig/... ./proxymode/...
+          go vet ./vpnservice/... ./persist/... ./proxyconfig/... ./proxymode/...
+          go test -race ./vpnservice/... ./persist/... ./proxyconfig/... ./proxymode/...
+          go build ./vpnservice/... ./persist/... ./proxyconfig/... ./proxymode/...
 
       # The desktop packaging script stamps the build version by setting
       # openrung/internal/client.appVersion with -X, but the linker silently
